Refined cluster deployment scripts.

+++ cluster/destroy_cluster_endpoint.sh
@@ -1,13 +1,29 @@
 #!/bin/bash
 
 set -eu
-
-BCM_CLUSTER_NAME=$1
-BCM_CLUSTER_ENDPOINT_NAME=$2
-
-# set the working directory to the location where the script is located
-# since all file references are relative to this script
 cd "$(dirname "$0")"
+
+BCM_CLUSTER_NAME=
+BCM_CLUSTER_ENDPOINT_NAME=
+
+for i in "$@"
+do
+case $i in
+    --cluster-name=*)
+    BCM_CLUSTER_NAME="${i#*=}"
+    shift # past argument=value
+    ;;
+    --endpoint-name=*)
+    BCM_CLUSTER_ENDPOINT_NAME="${i#*=}"
+    shift # past argument=value
+    ;;
+    *)
+
+    ;;
+esac
+done
+
+
 
 if [[ -z $BCM_CLUSTER_NAME ]]; then
   echo "BCM_CLUSTER_NAME not set. Exiting."
@@ -22,36 +38,45 @@ echo "BCM_CLUSTER_DIR: $BCM_CLUSTER_DIR"
 echo "ENDPOINTS_DIR: $ENDPOINTS_DIR"
 echo "BCM_ENDPOINT_DIR: $BCM_ENDPOINT_DIR"
 
-if [[ $(multipass list | grep "$BCM_CLUSTER_ENDPOINT_NAME") ]]; then
-  if [[ -f $BCM_ENDPOINT_DIR/.env ]]; then
-    source $BCM_ENDPOINT_DIR/.env
+
+function deleteLXDRemote {
+  # Removing lxc remote vm
+  if [[ $(lxc remote list | grep "$BCM_CLUSTER_ENDPOINT_NAME") ]]; then
+      echo "Removing lxd remote $BCM_CLUSTER_ENDPOINT_NAME"
+      lxc remote set-default local
+      lxc remote remove $BCM_CLUSTER_ENDPOINT_NAME
   fi
+}
+
+if [[ ! -f $BCM_ENDPOINT_DIR/.env ]]; then
+  echo "No $BCM_ENDPOINT_DIR/.env file exists to source. Cannot delete endpoint."
+  deleteLXDRemote
+  exit
 fi
 
-# quit if there are no multipass environment variables loaded.
+source $BCM_ENDPOINT_DIR/.env
+
+# Ensure the endpoint name is in our env.
 if [[ -z $(env | grep "$BCM_CLUSTER_ENDPOINT_NAME") ]]; then
   echo "BCM_CLUSTER_ENDPOINT_NAME variable not set."
   exit 1
 fi
 
-# Stopping multipass vm $MULTIPASS_VM_NAME
-if [[ $(multipass list | grep "$BCM_CLUSTER_ENDPOINT_NAME") ]]; then
-  echo "Stopping multipass vm $BCM_CLUSTER_ENDPOINT_NAME"
-  sudo multipass stop $BCM_CLUSTER_ENDPOINT_NAME
-  sudo multipass delete $BCM_CLUSTER_ENDPOINT_NAME
-  sudo multipass purge
-else
-  echo "$BCM_CLUSTER_ENDPOINT_NAME doesn't exist."
+if [[ $BCM_PROVIDER_NAME = "multipass" ]]; then
+  # Stopping multipass vm $MULTIPASS_VM_NAME
+  if [[ $(multipass list | grep "$BCM_CLUSTER_ENDPOINT_NAME") ]]; then
+    echo "Stopping multipass vm $BCM_CLUSTER_ENDPOINT_NAME"
+    sudo multipass stop $BCM_CLUSTER_ENDPOINT_NAME
+    sudo multipass delete $BCM_CLUSTER_ENDPOINT_NAME
+    sudo multipass purge
+  else
+    echo "$BCM_CLUSTER_ENDPOINT_NAME doesn't exist."
+  fi
+elif [[ $BCM_PROVIDER_NAME = "baremetal" ]]; then
+  echo "TODO baremetal destroy."
 fi
 
-# Removing lxc remote vm
-if [[ $(lxc remote list | grep $BCM_CLUSTER_ENDPOINT_NAME) ]]; then
-    echo "Removing lxd remote $BCM_CLUSTER_ENDPOINT_NAME"
-    lxc remote set-default local
-    lxc remote remove $BCM_CLUSTER_ENDPOINT_NAME
-else
-    echo "No lxc remote called $BCM_CLUSTER_ENDPOINT_NAME to delete."
-fi
+deleteLXDRemote
 
 if [[ -d $BCM_ENDPOINT_DIR ]]; then
   rm -rf $BCM_ENDPOINT_DIR
